Account for rotations when matching bodies (#3)
Currently, the addin matches bodies using their bounding boxes, which are always
axis-aligned. This fails to match rotated bodies, even for basic rectangular
shapes, which are pretty common. The ideal solution would be to find the minimal
bounding box of all shapes and use that, but the methods for computing minimal
3D bounding boxes seemed complicated to implement and expensive to run.

Instead, apply some heuristics that should work for most common woodworking
parts:

1. Find the planar face the largest perimeter
2. Determine the orientation of the face using the longest edge
3. Rotate the body so that the face normal is aligned with the Z axis and the
   orientation vector is aligned with the X axis
4. Use the bounding box of the rotated body

Finding the face orientation is the most likely place where this will fail,
particularly for faces with many curved edges.